+++ blog_user/src/main/java/cn/wangtao/controller/SysUserController.java
@@ -11,6 +11,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
@@ -39,16 +40,23 @@ public class SysUserController implements SysUserControllerApi {
         log.info("添加用户接受到参数：[{}]",user);
         BlogResponse blogResponse = new BlogResponse();
         HashMap<String, Object> map = new HashMap<>();
-        map.put("user",user);
-        SysUser sysUser = (SysUser)request.getSession().getAttribute("user");
-        if(sysUser!=null){
-            user.setCreateBy(sysUser.getUserName());//从Session中取
+        //参数校验
+        if(StringUtils.isEmpty(user.getUserName())){
+            blogResponse.setReturnCode(ConstantException.ERRORCODE);
+            blogResponse.setReturnMessage("增加失败，请求参数不合法");
+            log.error("增加失败，主键id不能为空");
+            return blogResponse;
+        }
+        SysUser currentUser = (SysUser)request.getSession().getAttribute("user");
+        if(currentUser!=null){
+            user.setCreateByName(currentUser.getUserName());//从Session中取
         }else{
             user.setSysRoleSeq(Constants.USER_COMMON);//普通用户
         }
 
         try{
             int num = sysUserService.insert(user);
+            map.put("user",user);
             blogResponse.setReturnCode(ConstantException.SUCCESSCODE);
             blogResponse.setReturnMessage("添加用户成功");
             blogResponse.setData(map);
@@ -56,7 +64,7 @@ public class SysUserController implements SysUserControllerApi {
         }catch (Exception e){
             blogResponse.setReturnCode(ConstantException.ERRORCODE);
             blogResponse.setReturnMessage("添加用户失败");
-            log.error("添加用户失败[{}]",user,e);
+            log.error("添加用户失败[{}]，，代码出现异常",user,e);
         }
         return blogResponse;
     }
@@ -77,7 +85,7 @@ public class SysUserController implements SysUserControllerApi {
         }catch (Exception e){
             blogResponse.setReturnCode(ConstantException.ERRORCODE);
             blogResponse.setReturnMessage("查询所有用户失败");
-            log.error("查询所有用户失败",e);
+            log.error("查询所有用户失败，代码出现异常",e);
         }
         return blogResponse;
     }
@@ -98,7 +106,7 @@ public class SysUserController implements SysUserControllerApi {
         }catch (Exception e){
             blogResponse.setReturnCode(ConstantException.ERRORCODE);
             blogResponse.setReturnMessage("通过属性字段[{}]查询用户失败");
-            log.error("通过属性字段[{}]查询用户失败",params,e);
+            log.error("通过属性字段[{}]查询用户失败，代码出现异常",params,e);
         }
         return blogResponse;
     }
@@ -109,6 +117,13 @@ public class SysUserController implements SysUserControllerApi {
         log.info("根据id查询对象接受到参数:[{}]",id);
         BlogResponse blogResponse = new BlogResponse();
         HashMap<String, Object> map = new HashMap<>();
+        //参数校验
+        if(StringUtils.isEmpty(id)){
+            blogResponse.setReturnCode(ConstantException.ERRORCODE);
+            blogResponse.setReturnMessage("查询失败，请求参数不合法");
+            log.error("查询失败，主键id不能为空");
+            return blogResponse;
+        }
         try {
             SysUser user = sysUserService.selectById(id);
             map.put("user",user);
@@ -119,7 +134,7 @@ public class SysUserController implements SysUserControllerApi {
         }catch (Exception e){
             blogResponse.setReturnCode(ConstantException.ERRORCODE);
             blogResponse.setReturnMessage("根据id查询对象失败");
-            log.error("根据id[{}]查询对象失败",id,e);
+            log.error("根据id[{}]查询对象失败，代码出现异常",id,e);
         }
         return blogResponse;
     }
@@ -129,31 +144,30 @@ public class SysUserController implements SysUserControllerApi {
     public BlogResponse updateById(SysUserModel userModel,HttpServletRequest request) {
         log.info("修改对象接受到参数:[{}]",userModel);
         BlogResponse blogResponse = new BlogResponse();
-        HashMap<String, Object> map = new HashMap<>();
         SysUser user =new SysUser();
+        //参数校验
+        if(StringUtils.isEmpty(userModel.getSysRoleSeq())){
+            blogResponse.setReturnCode(ConstantException.ERRORCODE);
+            blogResponse.setReturnMessage("修改失败，请求参数不合法");
+            log.error("修改失败，主键id不能为空");
+            return blogResponse;
+        }
+        //执行更新
         BeanUtils.copyProperties(userModel,user);
         //获取当前对象
         SysUser currentUser = (SysUser)request.getSession().getAttribute("user");
-        user.setUpdateBy(currentUser.getUserName());
+        if(currentUser!=null){
+            user.setUpdateBy(currentUser.getUserName());
+        }
         try {
             int num = sysUserService.update(user);
-            if(1==num){
-                SysUser sysUser=null;
-                try{
-                    sysUser= sysUserService.selectById(user.getSysUserSeq());
-                    log.info("修改对象成功[{}]记录数[{}]查询修改后的对象成功:[{}]",user,num,sysUser);
-                }catch (Exception e){
-                    log.error("修改对象成功[{}]记录数[{}]查询修改后的对象失败:[{}]",user,num,sysUser,e);
-                }
-                map.put("user",sysUser);
-                blogResponse.setReturnCode(ConstantException.SUCCESSCODE);
-                blogResponse.setReturnMessage("修改对象成功");
-                blogResponse.setData(map);
-            }
+            log.info("修改对象成功[{}]记录数[{}]",user,num);
+            blogResponse.setReturnCode(ConstantException.SUCCESSCODE);
+            blogResponse.setReturnMessage("修改对象成功");
         }catch (Exception e){
             blogResponse.setReturnCode(ConstantException.ERRORCODE);
             blogResponse.setReturnMessage("修改对象失败");
-            log.error("修改对象[{}]失败",user,e);
+            log.error("修改对象[{}]失败，代码出现异常",user,e);
         }
         return blogResponse;
     }
@@ -164,30 +178,42 @@ public class SysUserController implements SysUserControllerApi {
     public BlogResponse deleteById(@RequestParam("id") Long id,HttpServletRequest request) {
         log.info("根据id删除对象接受到参数:[{}]",id);
         BlogResponse blogResponse = new BlogResponse();
-        HashMap<String, Object> map = new HashMap<>();
+        //参数校验
+        if(StringUtils.isEmpty(id)){
+            blogResponse.setReturnCode(ConstantException.ERRORCODE);
+            blogResponse.setReturnMessage("删除失败，请求参数不合法");
+            log.error("删除失败，主键id不能为空");
+            return blogResponse;
+        }
         try {
             //获取当前对象
             SysUser currentUser = (SysUser)request.getSession().getAttribute("user");
             int num = sysUserService.deleteById(id, currentUser.getUserName());
             blogResponse.setReturnCode(ConstantException.SUCCESSCODE);
             blogResponse.setReturnMessage("根据id删除对象成功");
-            blogResponse.setData(map);
             log.info("根据id[{}]删除对象成功,记录数[{}]",id,num);
         }catch (Exception e){
             blogResponse.setReturnCode(ConstantException.ERRORCODE);
             blogResponse.setReturnMessage("根据id删除对象失败");
-            log.error("根据id[{}]删除对象失败",id,e);
+            log.error("根据id[{}]删除对象失败，代码出现异常",id,e);
         }
         return blogResponse;
     }
 
 
-    @PostMapping("selectByName")
+    @GetMapping("verifyName")
     @ResponseBody
-    public BlogResponse selectByName(String userName) {
+    public BlogResponse verifyName(String userName) {
         log.info("查询对象接受到参数:[{}]",userName);
         BlogResponse blogResponse = new BlogResponse();
         HashMap<String, Object> map = new HashMap<>();
+        //参数校验
+        if(StringUtils.isEmpty(userName)){
+            blogResponse.setReturnCode(ConstantException.ERRORCODE);
+            blogResponse.setReturnMessage("校验失败，请求参数不合法");
+            log.error("校验失败，主键id不能为空");
+            return blogResponse;
+        }
         try {
             SysUser user = sysUserService.selectByName(userName);
             if(user!=null){
@@ -203,9 +229,36 @@ public class SysUserController implements SysUserControllerApi {
         }catch (Exception e){
             blogResponse.setReturnCode(ConstantException.ERRORCODE);
             blogResponse.setReturnMessage("userName对象不存在");
-            log.error("根据userName[{}]查询对象失败",userName,e);
+            log.error("根据userName[{}]查询对象失败，代码出现异常",userName,e);
         }
         return blogResponse;
     }
 
+    @GetMapping("selectByName")
+    @ResponseBody
+    public BlogResponse selectByName(String userName) {
+        log.info("查询对象接受到参数:[{}]",userName);
+        BlogResponse blogResponse = new BlogResponse();
+        HashMap<String, Object> map = new HashMap<>();
+        //参数校验
+        if(StringUtils.isEmpty(userName)){
+            blogResponse.setReturnCode(ConstantException.ERRORCODE);
+            blogResponse.setReturnMessage("查询失败，请求参数不合法");
+            log.error("查询失败，userName参数为空");
+            return blogResponse;
+        }
+        try {
+            SysUser user = sysUserService.selectByName(userName);
+            map.put("user",user);
+            blogResponse.setReturnCode(ConstantException.SUCCESSCODE);
+            blogResponse.setReturnMessage("根据userName查询对象成功");
+            blogResponse.setData(map);
+            log.info("根据userName[{}]查询对象成功",userName);
+        }catch (Exception e){
+            blogResponse.setReturnCode(ConstantException.ERRORCODE);
+            blogResponse.setReturnMessage("userName对象不存在");
+            log.error("根据userName[{}]查询对象失败，代码出现异常",userName,e);
+        }
+        return blogResponse;
+    }
 }
